Новая версия, потому что старая не работала с float

+++ mathNotationTranslation/mathNotationTranslation/Test.c
@@ -27,7 +27,7 @@ static SExpression *allocateExpression()
 	return b;
 }
 
-SExpression *createNumber(int value)
+SExpression *createNumber(double value)
 {
 	SExpression *b = allocateExpression();
